streamlit/requirements.txt
streamlit==1.45.0
snowflake-connector-python==3.12.4
pandas==2.2.3
plotly==5.24.0
